+++ core/tauri-runtime/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## \[1.0.0-alpha.7]
+
+### Dependencies
+
+- Upgraded to `tauri-utils@2.0.0-alpha.12`
+
+## \[1.0.0-alpha.6]
+
+### Dependencies
+
+- [\`\`](https://www.github.com/tauri-apps/tauri/commit/undefined) Update dependencies.
+
 ## \[1.0.0-alpha.5]
 
 ### Dependencies